+++ frontend/src/pages/Home.jsx
@@ -1,0 +1,40 @@
+import { Container } from "@mui/material";
+import { useEffect, useState } from "react";
+import { Form } from "../components/Form";
+import { Timeline } from "../components/Timeline";
+
+export const Home = () => {
+  const [posts, setPosts] = useState([]);
+  const [isLoading, setIsLoading] = useState(false);
+  const onSubmitted = (post) => {
+    setPosts([post, ...posts]);
+  };
+
+  const fetchPosts = async () => {
+    setIsLoading(true);
+    const res = await fetch("/api/posts");
+    const data = await res.json();
+    if (!res.ok) {
+      console.error(data);
+      return;
+    }
+    setPosts(data);
+    setIsLoading(false);
+  };
+
+  useEffect(() => {
+    fetchPosts();
+  }, []);
+
+  return (
+    <Container
+      maxWidth="md"
+      sx={{
+        py: 3,
+      }}
+    >
+      <Form onSubmitted={onSubmitted} />
+      <Timeline posts={posts} isLoading={isLoading} fetchPosts={fetchPosts} />
+    </Container>
+  );
+};
